refactor(chat): extract text/artifact rendering into TextArtifactContentPart
